Reinstall Jellyfin (expected broken)
I'm going to check this in and push it, but not trigger a redeploy from
it, because it's won't work as-written - I don't think I activated
passing through the NVIDIA card from `epsilon` to the containers, so the
pod won't be able to satisfy these constraints. Will need to loosen them
until I can get around to setting up hardware acceleration - but good to
keep notes here in case needed.