Update architecture for Jellyfin
(As this matches the labels that are already applied on the nodes.
Previously, pod was not able to schedule)

I am _guessing_ that this means that the labelling convention has
updated in a version of k3s that I hadn't updated to?

Also perhaps need to update the toleration? Or not? My reading of
[docs](https://kubernetes.io/docs/concepts/scheduling-eviction/taint-and-toleration/)
is that it's not actually necessary (as only needed when they are used
as taints)